demo/pyproject.toml
[project]
name = "demo"
version = "0.1.0"
description = "Demo receiver for AI VR driver frames"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=1.24.0",
    "pygame-ce>=2.5.0",
]
